big one. time and weather work, removed some fonts.

+++ Core/UserLocation.cs
@@ -3,8 +3,8 @@
     public class UserLocation
     {
         public ulong Id { get; set; }
-        public float Lat { get; set; }
-        public float Lng { get; set; }
+        public double Lat { get; set; }
+        public double Lng { get; set; }
         public string City { get; set; }
         public int CityId { get; set; }
         public string TimeZone { get; set; }
